+++ packages/react-app/src/hooks/index.js
@@ -1,4 +1,4 @@
-export { default as useDebounce } from "./Debounce";
+export { default as useDebounce } from "./useDebounce";
 export { default as useLocalStorage } from "./useLocalStorage";
 export { default as useStaticJsonRPC } from "./useStaticJsonRPC";
 export * from "./useContractConfig";
